+++ state/state_machine.h
@@ -9,6 +9,7 @@
 #include <typeinfo>
 
 #include "static_str.h"
+#include "polymap.h"
 
 #ifndef __EMSCRIPTEN__
 #define CURR_STATE(SM) std::visit([](auto&&x) { \
@@ -26,33 +27,6 @@
 
 namespace Fresa::State
 {
-    //State Types
-    //-------------------------------------
-    template <typename... T>
-    struct Types {};
-
-    template <typename... L, typename... R>
-    constexpr auto operator+(Types<L...>, Types<R...>) {
-        return Types<L..., R...>{};
-    }
-
-    //: Cartesian Product (Domain of state function)
-    template <typename L, typename...R>
-    constexpr auto operator*(Types<L>, Types<R...>) {
-        return Types<Types<L, R>...>{};
-    }
-    template <typename... L, typename R>
-    constexpr auto operator*(Types<L...>, R r) {
-        return ((Types<L>{} * r) + ...);
-    }
-
-    //: Operate on every type
-    template <typename... T, typename Operation>
-    constexpr auto operator|(Types<T...>, Operation op) {
-        return op(Types<T>{}...);
-    }
-    //-------------------------------------
-
     //Apply operation and sum
     //-------------------------------------
     template <typename Operation>